Raytracer renders hit objects and general refactor

+++ RayTracer/Source/RayTracing/HitInfo.cs
@@ -6,7 +6,7 @@ namespace RayTracer.Maths
     public struct HitInfo
     {
         public Model ModelHit;
-        public double Distance;
+        public float Distance;
         public Vector3 HitPoint;
         public Vector3 Normal;
         public Vector2 TexCoord;
